Give Singapore a real affordability model instead of US underwriting (#33)
The Affordability tab ran the US DTI model in both regions, swapping only
the currency symbol. A Singapore buyer was asked for "max debt-to-income",
"monthly HOA" and a flat property-tax rate — none of which exist here —
and shown a confident S$458,187.04 derived from American underwriting.
That is worse than offering nothing, because it looks authoritative and is
precise to the cent.

Singapore runs on different rules, so this adds the model rather than
translating labels:

- Servicing capacity from the MAS TDSR (55%) and MSR (30%, HDB/EC only)
  ceilings, priced at the medium-term rate floor — so a 1.5% quote cannot
  buy more house than a 4% one, the same way a bank assesses it.

- LTV and minimum-cash rows from MAS Notice 632 (rev. 5 July 2018): 75%/5%
  for a first loan, 45%/25% for a second, 35%/25% for a third, each
  dropping to 55%/45%/35% -> 55/25/15% on extended tenure. Extended means
  over 30 years, or 25 for an HDB flat, or maturing past age 65. Rows are
  cited by their Notice reference in the code.

- BSD and ABSD charged on the resulting price, which makes the answer
  circular: the price a buyer can reach depends on duties levied on that
  price. Solved by bisection, since the funds a purchase needs are
  monotonic in price even though the progressive duty bands are not linear.

The panel names which rule actually binds. "You can afford $X" doesn't tell
a buyer whether to earn more or save more; "limited by your deposit, the
loan is already at the MAS LTV ceiling" does. Note this is deliberately the
constraint on the *loan*: a pricier property always needs a bigger deposit,
so funds bound the price for everyone and saying so tells nobody anything.

US affordability is untouched and now correctly scoped to US. The two are
separate components rather than one with branches, so neither carries the
other's fields.

Also fixes a float artifact at the boundary: converting a Decimal ratio and
then multiplying by 100.0 turned 0.55 into 55.00000000000001. Scaled in
Decimal and converted once. The existing SG panel had the same latent bug,
hidden only because it formats to two decimal places.

Verified live across four branches: baseline 75% LTV; 35-year tenure drops
to 55% and flags extended; an HDB flat switches the binding rule to MSR;
a foreigner's ABSD comes to exactly 60% of price and collapses the reachable
price from S$1.43M to S$461k. Deposit plus duties lands on the funds
available to the cent.

+++ crates/mortgage-wasm/src/singapore.rs
@@ -53,8 +53,12 @@ fn parse_property_count(s: &str) -> singapore::PropertyCount {
 
 /// Ratios come back from `mortgage_calc` as fractions (0.482); the UI shows
 /// percentages.
+///
+/// Scaled in `Decimal` and converted once. Converting first and then
+/// multiplying by 100.0 lands on values like 55.00000000000001, which
+/// survives into any caller that doesn't round on the way out.
 fn to_percent(ratio: Decimal) -> f64 {
-    decimal_to_f64(ratio) * 100.0
+    decimal_to_f64(ratio * Decimal::from(100))
 }
 
 /// The JsValue-free core, so it can be unit-tested with plain
